requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
PyPDF2
numpy
imaplib-status
python-email-validator